Cleanup, error handling refactoring, naming convention.
- Fully use SurfTimer.Shared entities
- Delete unused files and methods
- Get rid of some comments and messages
- Rename classes to follow PascalCase
- Better(?) error handling
- Change types to match with DB
- Start using `timer_settings.json` (does NOT stop replays atm)

+++ src/ST-Player/PlayerTimer.cs
@@ -13,11 +13,10 @@ public class PlayerTimer
     public bool IsBonusMode { get; set; } = false; // Bonus mode toggle
 
     // Tracking
-    public int Stage { get; set; } = 0; // Current stage tracker
-    public int Checkpoint {get; set;} = 0; // Current checkpoint tracker
-    // public CurrentRun CurrentRunData { get; set; } = new CurrentRun(); // Current RUN data tracker
-    public int Bonus { get; set; } = 0; // To-do: bonus implementation - Current bonus tracker 
-    public int Style { get; set; } = 0; // To-do: functionality for player to change this value and the actual styles implementation - Current style tracker
+    public short Stage { get; set; } = 0; // Current stage tracker
+    public short Checkpoint { get; set; } = 0; // Current checkpoint tracker
+    public short Bonus { get; set; } = 0; // To-do: bonus implementation - Current bonus tracker 
+    public short Style { get; set; } = 0; // To-do: functionality for player to change this value and the actual styles implementation - Current style tracker
 
     // Timing
     public int Ticks { get; set; } = 0; // To-do: sub-tick counting? This currently goes on OnTick, which is not sub-tick I believe? Needs investigating
@@ -44,7 +43,6 @@ public class PlayerTimer
         this.IsPracticeMode = false;
         this.IsStageMode = false;
         this.IsBonusMode = false;
-        // this.CurrentRunData.Reset();
     }
 
     public void Pause()
